requirements.txt
# Core dependencies
playwright==1.38.0
python-dotenv==1.0.0

# Utilities
pytest==7.4.0
pytest-playwright==0.4.0
pytest-html==3.2.0
Appium-Python-Client==3.1.0